Allegrex binutils for v2.37
This is a beta relase, some minor issues still present.

Should be compatible with the older 2.23-based version except for the
syntax of vpfxs/t/d instructions, that now require brackets `[]` around
the arguments. This was necessary to simplify the assembler since the
parter is now generic (tokens are parsed independently of the opcode).

Other instructions such as vrot could have different (yet equivalent)
binary encoding, since they have ambiguous encodings. I've tried to make
the binary encoding as compatible as possible with the previous version,
so that automated testing can be used to compare them.

There might be bugs or missing checks on things such as register
collision logic, unallowed prefixes and similar stuff. However code that
worked with the previous version should still work.