Allegrex binutils for v2.37
This the v2 release candidate patch for binutils v2.37.

Should be compatible with the older 2.23-based version except for the
syntax of vpfxs/t/d instructions, that now require brackets `[]` around
the arguments. This was necessary to simplify the assembler since the
parter is now generic (tokens are parsed independently of the opcode).

Other instructions such as vrot could have different (yet equivalent)
binary encoding, since they have ambiguous encodings. I've tried to make
the binary encoding as compatible as possible with the previous version,
so that automated testing can be used to compare them.

There might be bugs or missing checks on things such as register
collision logic, unallowed prefixes and similar stuff. However code that
worked with the previous version should still work.

Register names in VFPU functions could collide with variable names if
they are identical (like in lv/sv instructions). That's an issue with
the original register naming (it should have a '$' symbol as a prefix
but it doesn't). Anyway v2 should really fix collisions with other
constants and condition codes.